Split the property parsing logic into its own service
